requirement.txt
newsapi-python==0.2.6
numpy==1.19.1
pandas==1.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.12.5
six==1.15.0
